feat: promisify in place
