+++ src/batch.js
@@ -2,7 +2,6 @@
 
 'use strict';
 
-var async = require('async');
 const util = require('util');
 
 var db = require('./database');
@@ -42,26 +41,22 @@ exports.processSortedSet = async function (setKey, process, options) {
 		process = util.promisify(process);
 	}
 
-	await async.whilst(
-		function (next) {
-			next(null, !done);
-		},
-		async function () {
-			const ids = await db['getSortedSetRange' + (options.withScores ? 'WithScores' : '')](setKey, start, stop);
-			if (!ids.length || options.doneIf(start, stop, ids)) {
-				done = true;
-				return;
-			}
-			await process(ids);
-
-			start += utils.isNumber(options.alwaysStartAt) ? options.alwaysStartAt : options.batch + 1;
-			stop = start + options.batch;
-
-			if (options.interval) {
-				await sleep(options.interval);
-			}
+	while (!done) {
+		/* eslint-disable no-await-in-loop */
+		const ids = await db['getSortedSetRange' + (options.withScores ? 'WithScores' : '')](setKey, start, stop);
+		if (!ids.length || options.doneIf(start, stop, ids)) {
+			done = true;
+			return;
 		}
-	);
+		await process(ids);
+
+		start += utils.isNumber(options.alwaysStartAt) ? options.alwaysStartAt : options.batch + 1;
+		stop = start + options.batch;
+
+		if (options.interval) {
+			await sleep(options.interval);
+		}
+	}
 };
 
 exports.processArray = async function (array, process, options) {
@@ -80,23 +75,22 @@ exports.processArray = async function (array, process, options) {
 	if (process && process.constructor && process.constructor.name !== 'AsyncFunction') {
 		process = util.promisify(process);
 	}
-	await async.whilst(
-		function (next) {
-			next(null, !done);
-		},
-		async function () {
-			var currentBatch = array.slice(start, start + batch);
-			if (!currentBatch.length) {
-				done = true;
-				return;
-			}
-			await process(currentBatch);
-			start += batch;
-			if (options.interval) {
-				await sleep(options.interval);
-			}
-		}
-	);
-};
 
-require('./callbackify')(exports);
+	while (!done) {
+		var currentBatch = array.slice(start, start + batch);
+
+		if (!currentBatch.length) {
+			done = true;
+			break;
+		}
+
+		await process(currentBatch);
+
+		start += batch;
+
+		if (options.interval) {
+			await sleep(options.interval);
+		}
+	}
+};
+require('./promisify')(exports);
