+++ src/database/mongo.js
@@ -127,7 +127,6 @@ mongoModule.init = function (callback) {
 		require('./mongo/transaction')(db, mongoModule);
 
 		mongoModule.async = require('../promisify')(mongoModule, ['client', 'sessionStore']);
-		require('../callbackify')(mongoModule, ['client', 'sessionStore']);
 		callback();
 	});
 };
